ETA to next waypoint and steer error (#77)
* Create eta.js

I would like to add 'eta' to next waypoint to derived data.  this file creates:
'navigation.courseGreatCircle.nextPoint.eta' and delivers a date/time of arrival at the waypoint

please let me know if I need to do anything else to add it?

* Create steer_error.js

this file adds a few phrases for Steer Error, this being the difference between the desired direction and the actual direction.  there are 3 different phrases:

"navigation.courseGreatCircle.nextPoint.steerError" - this gives a value for the error with a positive or negative to suggest the direction
"navigation.courseGreatCircle.nextPoint.leftSteerError" - this gives a value for the error in the Port direction
"navigation.courseGreatCircle.nextPoint.rightSteerError" - this gives a value for the error in the Starboard direction
the right and left steer error are for use to help with clear race displays to help a crew member get clear direction in the heat of a race