fix: notification clearing
The code was sending the normal state notification to the
other vessel, when in fact the context must be self,
where the original notification was raised.

Use an object, not array, to keep track of raised
notifications. The code was a little confused in this
respect.

Remove the length < 1 condition from clearing the
notifications on stop, as it makes no sense to check
for 0 length array and then iterate over it.